Creo que quedo todo el usuario

+++ src/main/java/mx/edu/utex/APREHO/services/ServicesUser/UserService.java
@@ -30,6 +30,10 @@ public class UserService {
     @Transactional(rollbackFor = {SQLException.class})
     public ResponseEntity<ApiResponse> save(User user) {
 
+        Optional<User> foundUser=repository.findByEmail(user.getEmail());
+        if(foundUser.isPresent()){
+            return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST, true, "Este correo ya se ha registrado previamente"), HttpStatus.BAD_REQUEST);
+        }else{
             if (user.getPeople() != null) {
                 Optional<People> foundPeople = peopleRepository.findByCurp(user.getPeople().getCurp());
                 if (!foundPeople.isPresent()) {
@@ -53,7 +57,7 @@ public class UserService {
                 return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST, true, "No has ingresado a un ROL para este Usuario"), HttpStatus.BAD_REQUEST);
 
             }
-
+        }
 
         return new ResponseEntity<>(new ApiResponse(repository.saveAndFlush(user), HttpStatus.OK, false, "usuario creado exitosamente"), HttpStatus.OK);
     }
@@ -97,24 +101,53 @@ public class UserService {
     @Transactional(rollbackFor = {SQLException.class})
     public ResponseEntity<ApiResponse> delete(User user) {
 
+        Optional<User> foundUser = repository.findById(user.getUserId());
+        if (foundUser.isPresent()) {
             if (user.getPeople() != null) {
-                Optional<People> foundPeople = peopleRepository.findByCurp(user.getPeople().getCurp());
+                Optional<People> foundPeople = peopleRepository.findById(user.getPeople().getPeopleId());
                 if (foundPeople.isPresent()) {
                     People people = foundPeople.get();
                     peopleRepository.delete(people);
-                    repository.delete(user);
-                    return new ResponseEntity<>(new ApiResponse( HttpStatus.OK, false, "usuario Eliminado exitosamente"), HttpStatus.OK);
+                    User user1 = foundUser.get();
+                    repository.delete(user1);
+                    return new ResponseEntity<>(new ApiResponse(HttpStatus.OK, false, "usuario Eliminado exitosamente"), HttpStatus.OK);
 
-                }else {
+                } else {
                     return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST, true, "No hemos encontrado a la persona"), HttpStatus.BAD_REQUEST);
 
                 }
-            }else {
-                return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST, true, "Este usuario no existe"), HttpStatus.BAD_REQUEST);
+            } else {
+                return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST, true, "Esta persona no existe"), HttpStatus.BAD_REQUEST);
 
             }
+        }
 
+        return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST, true, "Este usuario no existe"), HttpStatus.BAD_REQUEST);
+    }
 
+    @Transactional(rollbackFor = {SQLException.class})
+    public  ResponseEntity<ApiResponse> findOne(Long id){
+        Optional<User> foundUser=repository.findById(id);
+        if (foundUser.isPresent()){
+            return new ResponseEntity<>(new ApiResponse(repository.findOne(id),HttpStatus.OK,false,"Usuario encontrado"),HttpStatus.OK);
+
+        }else{
+            return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST,false,"Usuario no encontrado"),HttpStatus.BAD_REQUEST);
+
+        }
+
+    }
+
+    @Transactional(rollbackFor = {SQLException.class})
+    public  ResponseEntity<ApiResponse> loggin(String pass,String email){
+        Optional<User> foundUser=repository.loggin(pass,email);
+        if(foundUser.isPresent()){
+            return new ResponseEntity<>(new ApiResponse(foundUser,HttpStatus.OK,false,"Usuario encontrado"),HttpStatus.OK);
+
+        }else{
+            return new ResponseEntity<>(new ApiResponse(HttpStatus.BAD_REQUEST,false,"Usuario o contraseña Incorrectas"),HttpStatus.BAD_REQUEST);
+
+        }
     }
 
 }
